Dex's suggestion to maybe stop the negative length crashs

+++ src/main/java/pcl/bridgebot/chatserverlink/ChatServerLink.java
@@ -107,7 +107,7 @@ public class ChatServerLink {
 				int dataRead = inFromServer.read(buffer, 0, 2);
 				if (dataRead != 4) {
 					// Manually read a LE short from the stream
-					short packetSize = (short)((((short)(buffer[1])) << 8) + ((short) (buffer[0])));
+					int packetSize = (int)((((int)(buffer[1])) << 8) + ((int) (buffer[0])));
 					if(packetSize > 1000) {
 						System.out.println(String.format("Announced size was 0x%02X bytes, longer than the buffer. Skipping packet", packetSize));
 						continue;
